1.实现净值更新 2.实现刷新

+++ Simple/Controllers/YJEditViewController.swift
@@ -40,7 +40,7 @@ class YJEditViewController: UIViewController,UITextFieldDelegate,UISearchBarDele
             nameField.becomeFirstResponder()
             return
         }
-        guard let amount:Int64 = Int64(amountField.text!)else{
+        guard let amount:Double = Double(amountField.text!)else{
             amountField.becomeFirstResponder()
             return
         }
